* put testcase into main_test.go
